Move VmdbDatabaseConnection specific code out of MiqDbConfig.
