+++ app/models/miq_db_config.rb
@@ -198,32 +198,6 @@ class MiqDbConfig
     false
   end
 
-  def self.log_statistics
-    log_activity_statistics
-  end
-
-  def self.log_activity_statistics(output = $log)
-    require 'csv'
-
-    begin
-      stats = VmdbDatabaseConnection.all.map(&:to_csv_hash)
-
-      keys = stats.first.keys
-
-      csv = CSV.generate do |rows|
-        rows << keys
-        stats.each do |s|
-          vals = s.values_at(*keys)
-          rows << vals
-        end
-      end
-
-      output.info("MIQ(DbConfig.log_activity_statistics) <<-ACTIVITY_STATS_CSV\n#{csv}ACTIVITY_STATS_CSV")
-    rescue => err
-      output.warn("MIQ(DbConfig.log_activity_statistics) Unable to log stats, '#{err.message}'")
-    end
-  end
-
   private
 
   def with_temporary_connection
