+++ app/models/vmdb_database_connection.rb
@@ -32,6 +32,28 @@ class VmdbDatabaseConnection < ApplicationRecord
     false
   end
 
+  def self.log_statistics(output = $log)
+    require 'csv'
+
+    begin
+      stats = all.map(&:to_csv_hash)
+
+      keys = stats.first.keys
+
+      csv = CSV.generate do |rows|
+        rows << keys
+        stats.each do |s|
+          vals = s.values_at(*keys)
+          rows << vals
+        end
+      end
+
+      output.info("MIQ(#{name}.#{__method__}) <<-ACTIVITY_STATS_CSV\n#{csv}ACTIVITY_STATS_CSV")
+    rescue => err
+      output.warn("MIQ(#{name}.#{__method__}) Unable to log stats, '#{err.message}'")
+    end
+  end
+
   def vmdb_database_id
     @vmdb_database_id ||= self.class.vmdb_database.id
   end
